Refactored index as a client, added OEIS.js as the actual module
